refactor(gui): P4 review fixups — extract highlight cache + side pane
Addresses HIGH + MEDIUM findings from rust-reviewer and code-reviewer
on commit 832bfeb:

- Extract highlight_cache.rs (side_lines, highlight_by_lineno,
  side_cache_key_for_text) and side_pane.rs (SideRow, SBS builders)
  so mod.rs (1456 -> 710) and diff_pane.rs (847 -> 573) drop below
  the 800-line soft limit.
- Drop DefaultHasher from the content fallback key: use the raw
  content string with a length prefix so the key is deterministic
  across processes and free of collision risk for caches that may
  later outlive a single run.
- Unify cache keying via side_cache_key_for_text(file, side, syntax,
  text) used by BOTH maybe_spawn_highlight (storage) and
  push_selected_file_highlights (lookup). Avoids the double
  side_lines hunk-walk per working-tree diff.
- HighlightEngine::global now returns a static reference matching
  the RFC 8.4 signature (no Arc wrapper). Add HighlightEngine::prime
  as the explicit warm-up call in DiffView::new.
- Rename _highlighter to highlighter_task; drop the clippy allow
  attributes on init/cancel/spawn sites.
- Tests: 44 passed (+9 net). New cache_key_storage_and_lookup_paths_
  match guards future divergence between sides; new
  content_fallback_key_is_deterministic_and_distinct replaces the
  old hash test.
- RFC: document ViewMode persistence as a Phase 6 deferral in a new
  Deviations section.
- Theme: note that syntax_*() are currently dark-only; light-theme
  variants land when a light theme does.
