+++ crates/zremote-gui/src/views/diff/mod.rs
@@ -9,7 +9,9 @@
 pub mod diff_pane;
 pub mod file_tree;
 pub mod highlight;
+pub mod highlight_cache;
 pub mod large_file;
+pub mod side_pane;
 pub mod source_picker;
 pub mod state;
 
@@ -20,7 +22,7 @@ use futures_util::StreamExt;
 use gpui::*;
 
 use zremote_client::diff::{DiffEventWire, get_diff_sources, stream_diff};
-use zremote_protocol::project::{DiffFile, DiffLineKind, DiffRequest, DiffSource};
+use zremote_protocol::project::{DiffFile, DiffRequest, DiffSource};
 
 use crate::app_state::AppState;
 use crate::icons::{Icon, icon};
@@ -28,7 +30,8 @@ use crate::theme;
 
 use self::diff_pane::{DiffPane, HighlightCache};
 use self::file_tree::{FileTree, FileTreeEvent};
-use self::highlight::{HighlightEngine, LineSpans, SideKey, should_highlight};
+use self::highlight::{HighlightEngine, SideKey, should_highlight};
+use self::highlight_cache::{highlight_by_lineno, side_cache_key_for_text, side_lines};
 use self::source_picker::{SourcePicker, SourcePickerEvent};
 use self::state::{DiffEvent, DiffState, ViewMode, apply};
 
@@ -56,7 +59,7 @@ pub struct DiffView {
     /// Background syntax-highlighting task for the currently selected
     /// file. Replacing this field drops (and cancels) the previous task,
     /// which is important when the user navigates file → file quickly.
-    _highlighter: Option<Task<()>>,
+    highlighter_task: Option<Task<()>>,
     /// Per-file, per-side highlight cache. Key is
     /// `(blob_sha_or_content_hash, syntax_name, SideKey)`. Cleared on
     /// project switch / source change so old hunks don't bleed into a new
@@ -86,7 +89,7 @@ impl DiffView {
         // caches the result for every subsequent synchronous `global()`
         // call from render paths.
         cx.background_spawn(async {
-            let _ = HighlightEngine::global();
+            HighlightEngine::prime();
         })
         .detach();
 
@@ -125,7 +128,7 @@ impl DiffView {
             diff_pane,
             stream_task: None,
             sources_task: None,
-            _highlighter: None,
+            highlighter_task: None,
             highlight_cache: HashMap::new(),
             _child_subs: subs,
         };
@@ -159,13 +162,9 @@ impl DiffView {
 
         if clear_highlight_cache {
             self.highlight_cache.clear();
-            // Drop any in-flight highlight work keyed to the old source.
-            // Underscore-prefix is intentional (RFC §8.2 convention: fields
-            // whose only purpose is "drop cancels the task" read as _field).
-            #[allow(clippy::used_underscore_binding)]
-            {
-                self._highlighter = None;
-            }
+            // Drop any in-flight highlight work keyed to the old source:
+            // replacing the Task cancels the previous one (RFC §8.2).
+            self.highlighter_task = None;
         }
         if update_view_mode {
             let mode = self.state.view_mode;
@@ -241,12 +240,16 @@ impl DiffView {
             .detect_syntax(&file.summary.path)
             .name
             .clone();
-        // Use the same fallback logic as `maybe_spawn_highlight` so storage
-        // and lookup keys always agree. Working-tree diffs have `new_sha =
-        // None` (see `parser.rs`); without this fallback, highlights were
-        // computed + cached but never delivered to the pane.
-        let old_key = side_cache_key(file, SideKey::Old, &syntax_name);
-        let new_key = side_cache_key(file, SideKey::New, &syntax_name);
+        // Derive keys from the same `side_lines` output that the storage
+        // path uses. Both paths must call `side_cache_key_for_text` with
+        // identical text, otherwise highlights get stored under one key
+        // and searched under another.
+        let old_lines = side_lines(file, SideKey::Old);
+        let new_lines = side_lines(file, SideKey::New);
+        let old_text: String = old_lines.iter().map(|(_, t)| t.as_str()).collect();
+        let new_text: String = new_lines.iter().map(|(_, t)| t.as_str()).collect();
+        let old_key = side_cache_key_for_text(file, SideKey::Old, &syntax_name, &old_text);
+        let new_key = side_cache_key_for_text(file, SideKey::New, &syntax_name, &new_text);
         let old = self.highlight_cache.get(&old_key).cloned();
         let new = self.highlight_cache.get(&new_key).cloned();
         self.diff_pane.update(cx, |pane, cx| {
@@ -280,10 +283,13 @@ impl DiffView {
         let new_lines = side_lines(file, SideKey::New);
 
         let path = file.summary.path.clone();
-        let old_key = side_cache_key(file, SideKey::Old, &syntax_name);
-        let new_key = side_cache_key(file, SideKey::New, &syntax_name);
         let old_text: String = old_lines.iter().map(|(_, t)| t.as_str()).collect();
         let new_text: String = new_lines.iter().map(|(_, t)| t.as_str()).collect();
+        // Derive cache keys from the already-computed text rather than
+        // re-walking hunks inside a helper. This is what ties storage and
+        // lookup paths to a single key derivation (`side_cache_key_for_text`).
+        let old_key = side_cache_key_for_text(file, SideKey::Old, &syntax_name, &old_text);
+        let new_key = side_cache_key_for_text(file, SideKey::New, &syntax_name, &new_text);
         let needs_old = !self.highlight_cache.contains_key(&old_key)
             && !old_text.is_empty()
             && should_highlight(&old_text);
@@ -296,53 +302,50 @@ impl DiffView {
         }
 
         let selected_path = path.clone();
-        #[allow(clippy::used_underscore_binding)]
-        {
-            self._highlighter = Some(cx.spawn(async move |this, cx| {
-                // Run the pure-CPU highlight work off the GPUI thread so a
-                // slow syntect pass can't stall frame rendering.
-                let result = cx
-                    .background_spawn(async move {
-                        let engine = HighlightEngine::global();
-                        let syntax = engine.detect_syntax(&path);
-                        let old = if needs_old {
-                            Some((
-                                old_key.clone(),
-                                Arc::new(highlight_by_lineno(engine, syntax, &old_lines)),
-                            ))
-                        } else {
-                            None
-                        };
-                        let new = if needs_new {
-                            Some((
-                                new_key.clone(),
-                                Arc::new(highlight_by_lineno(engine, syntax, &new_lines)),
-                            ))
-                        } else {
-                            None
-                        };
-                        (old, new)
-                    })
-                    .await;
+        self.highlighter_task = Some(cx.spawn(async move |this, cx| {
+            // Run the pure-CPU highlight work off the GPUI thread so a
+            // slow syntect pass can't stall frame rendering.
+            let result = cx
+                .background_spawn(async move {
+                    let engine = HighlightEngine::global();
+                    let syntax = engine.detect_syntax(&path);
+                    let old = if needs_old {
+                        Some((
+                            old_key.clone(),
+                            Arc::new(highlight_by_lineno(engine, syntax, &old_lines)),
+                        ))
+                    } else {
+                        None
+                    };
+                    let new = if needs_new {
+                        Some((
+                            new_key.clone(),
+                            Arc::new(highlight_by_lineno(engine, syntax, &new_lines)),
+                        ))
+                    } else {
+                        None
+                    };
+                    (old, new)
+                })
+                .await;
 
-                let (old, new) = result;
-                let _ = this.update(cx, |this, cx| {
-                    // Storage is a lineno→spans map so multi-hunk files
-                    // highlight correctly (no padding, no positional
-                    // indexing from the start of a concatenated side).
-                    if let Some((key, spans)) = old {
-                        this.highlight_cache.insert(key, spans);
-                    }
-                    if let Some((key, spans)) = new {
-                        this.highlight_cache.insert(key, spans);
-                    }
-                    // Only push if the user is still on the same file.
-                    if this.state.selected_file.as_deref() == Some(&selected_path) {
-                        this.push_selected_file_highlights(cx);
-                    }
-                });
-            }));
-        }
+            let (old, new) = result;
+            let _ = this.update(cx, |this, cx| {
+                // Storage is a lineno→spans map so multi-hunk files
+                // highlight correctly (no padding, no positional
+                // indexing from the start of a concatenated side).
+                if let Some((key, spans)) = old {
+                    this.highlight_cache.insert(key, spans);
+                }
+                if let Some((key, spans)) = new {
+                    this.highlight_cache.insert(key, spans);
+                }
+                // Only push if the user is still on the same file.
+                if this.state.selected_file.as_deref() == Some(&selected_path) {
+                    this.push_selected_file_highlights(cx);
+                }
+            });
+        }));
     }
 
     /// Request that this diff view be closed. Emitted to the parent
@@ -686,102 +689,6 @@ impl DiffView {
     }
 }
 
-/// Collect the lines on one side of a diff (pre- or post-image) together
-/// with their 1-based line numbers. Highlighter fires one syntect pass over
-/// the concatenated text but stores results keyed by the real file line
-/// number, so multi-hunk files keep correct spans for every hunk (not just
-/// the first).
-///
-/// "Relevant lines" means:
-/// - `SideKey::Old`: `Context` + `Removed` lines (anything with an
-///   `old_lineno`).
-/// - `SideKey::New`: `Context` + `Added` lines (anything with a
-///   `new_lineno`).
-///
-/// Each returned text ends in a newline so syntect's state machine closes
-/// each line cleanly.
-fn side_lines(file: &DiffFile, side: SideKey) -> Vec<(u32, String)> {
-    let mut out = Vec::new();
-    for hunk in &file.hunks {
-        for line in &hunk.lines {
-            let include = match side {
-                SideKey::Old => {
-                    matches!(line.kind, DiffLineKind::Context | DiffLineKind::Removed)
-                }
-                SideKey::New => {
-                    matches!(line.kind, DiffLineKind::Context | DiffLineKind::Added)
-                }
-            };
-            if !include {
-                continue;
-            }
-            let lineno = match side {
-                SideKey::Old => line.old_lineno,
-                SideKey::New => line.new_lineno,
-            };
-            let Some(lineno) = lineno else {
-                continue;
-            };
-            let mut text = line.content.clone();
-            if !text.ends_with('\n') {
-                text.push('\n');
-            }
-            out.push((lineno, text));
-        }
-    }
-    out
-}
-
-/// Run syntect over the concatenated side text and distribute per-line
-/// spans into a `lineno → spans` map. The cache is indexed by real 1-based
-/// file line number (never by position in the concatenated string) so
-/// multi-hunk files produce correct lookups for every hunk.
-fn highlight_by_lineno(
-    engine: &HighlightEngine,
-    syntax: &syntect::parsing::SyntaxReference,
-    lines: &[(u32, String)],
-) -> HashMap<u32, LineSpans> {
-    let text: String = lines.iter().map(|(_, t)| t.as_str()).collect();
-    let spans = engine.highlight_file(&text, syntax);
-    let mut map = HashMap::with_capacity(lines.len());
-    // `highlight_file` emits one entry per source line in `text`, so the
-    // index order matches `lines`. If syntect ever returns fewer entries
-    // (pathological input), we silently cap the iteration — missing spans
-    // just leave those lines unhighlighted.
-    for ((lineno, _), line_spans) in lines.iter().zip(spans.into_iter()) {
-        map.insert(*lineno, line_spans);
-    }
-    map
-}
-
-/// Cache-key fallback for files whose protocol blob SHA is absent
-/// (working-tree diffs on a repo with no initial commit, for example).
-/// A short hash over the content is stable enough — collisions across
-/// unrelated files would still land on disjoint `(syntax_name, SideKey)`
-/// pairs because both are part of the full cache key.
-fn content_fallback_key(text: &str) -> String {
-    use std::hash::{Hash, Hasher};
-    let mut hasher = std::collections::hash_map::DefaultHasher::new();
-    text.hash(&mut hasher);
-    format!("content:{:x}", hasher.finish())
-}
-
-/// Canonical cache key for a file side. Falls back to a content hash when
-/// the protocol omitted a blob SHA (working-tree diffs always do). Storage
-/// and lookup paths must share this helper — otherwise highlights get
-/// stored under one key and searched under another.
-fn side_cache_key(file: &DiffFile, side: SideKey, syntax_name: &str) -> (String, String, SideKey) {
-    let sha = match side {
-        SideKey::Old => file.summary.old_sha.as_ref(),
-        SideKey::New => file.summary.new_sha.as_ref(),
-    };
-    let sha_or_fallback = sha.cloned().unwrap_or_else(|| {
-        let text: String = side_lines(file, side).into_iter().map(|(_, t)| t).collect();
-        content_fallback_key(&text)
-    });
-    (sha_or_fallback, syntax_name.to_string(), side)
-}
-
 /// Local text tooltip used by the close button. Duplicated intentionally from
 /// `sidebar::SidebarTextTooltip` since that type is private to the sidebar
 /// module; a shared component would require wider refactoring out of P3 scope.
@@ -799,196 +706,5 @@ impl Render for DiffTextTooltip {
             .text_size(px(11.0))
             .text_color(theme::text_secondary())
             .child(self.0.clone())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{
-        HighlightEngine, SideKey, content_fallback_key, highlight_by_lineno, side_cache_key,
-        side_lines,
-    };
-    use zremote_protocol::project::{
-        DiffFile, DiffFileStatus, DiffFileSummary, DiffHunk, DiffLine, DiffLineKind,
-    };
-
-    fn mk_line(kind: DiffLineKind, old: Option<u32>, new: Option<u32>, content: &str) -> DiffLine {
-        DiffLine {
-            kind,
-            old_lineno: old,
-            new_lineno: new,
-            content: content.to_string(),
-        }
-    }
-
-    fn mk_file(
-        path: &str,
-        old_sha: Option<&str>,
-        new_sha: Option<&str>,
-        hunks: Vec<DiffHunk>,
-    ) -> DiffFile {
-        DiffFile {
-            summary: DiffFileSummary {
-                path: path.to_string(),
-                old_path: None,
-                status: DiffFileStatus::Modified,
-                binary: false,
-                submodule: false,
-                too_large: false,
-                additions: 0,
-                deletions: 0,
-                old_sha: old_sha.map(String::from),
-                new_sha: new_sha.map(String::from),
-                old_mode: None,
-                new_mode: None,
-            },
-            hunks,
-        }
-    }
-
-    #[test]
-    fn side_lines_preserves_real_linenos_across_multiple_hunks() {
-        // hunk 1 starts at line 10; hunk 2 starts at line 50. Storage must
-        // remember the real line numbers, not position-in-concatenated-text.
-        let hunks = vec![
-            DiffHunk {
-                old_start: 10,
-                old_lines: 3,
-                new_start: 10,
-                new_lines: 3,
-                header: "@@ -10,3 +10,3 @@".into(),
-                lines: vec![
-                    mk_line(DiffLineKind::Context, Some(10), Some(10), "fn a() {\n"),
-                    mk_line(DiffLineKind::Context, Some(11), Some(11), "    body\n"),
-                    mk_line(DiffLineKind::Context, Some(12), Some(12), "}\n"),
-                ],
-            },
-            DiffHunk {
-                old_start: 50,
-                old_lines: 3,
-                new_start: 50,
-                new_lines: 3,
-                header: "@@ -50,3 +50,3 @@".into(),
-                lines: vec![
-                    mk_line(DiffLineKind::Context, Some(50), Some(50), "fn b() {\n"),
-                    mk_line(DiffLineKind::Context, Some(51), Some(51), "    body2\n"),
-                    mk_line(DiffLineKind::Context, Some(52), Some(52), "}\n"),
-                ],
-            },
-        ];
-        let file = mk_file("x.rs", Some("old"), Some("new"), hunks);
-        let lines = side_lines(&file, SideKey::New);
-        let linenos: Vec<u32> = lines.iter().map(|(n, _)| *n).collect();
-        assert_eq!(linenos, vec![10, 11, 12, 50, 51, 52]);
-    }
-
-    #[test]
-    fn highlights_available_for_all_hunks_in_multi_hunk_file() {
-        // Regression guard for B1: pre-fix storage used positional
-        // padding, so hunk 2 starting at line 50 fell off the end of the
-        // span vector. With the lineno-indexed HashMap this test asserts
-        // lookup works for BOTH hunks.
-        let hunks = vec![
-            DiffHunk {
-                old_start: 10,
-                old_lines: 2,
-                new_start: 10,
-                new_lines: 2,
-                header: "@@ -10,2 +10,2 @@".into(),
-                lines: vec![
-                    mk_line(DiffLineKind::Context, Some(10), Some(10), "let a = 1;\n"),
-                    mk_line(DiffLineKind::Context, Some(11), Some(11), "let b = 2;\n"),
-                ],
-            },
-            DiffHunk {
-                old_start: 50,
-                old_lines: 2,
-                new_start: 50,
-                new_lines: 2,
-                header: "@@ -50,2 +50,2 @@".into(),
-                lines: vec![
-                    mk_line(DiffLineKind::Context, Some(50), Some(50), "let c = 3;\n"),
-                    mk_line(DiffLineKind::Context, Some(51), Some(51), "let d = 4;\n"),
-                ],
-            },
-        ];
-        let file = mk_file("x.rs", Some("old"), Some("new"), hunks);
-        let engine = HighlightEngine::global();
-        let syntax = engine.detect_syntax("x.rs");
-        let lines = side_lines(&file, SideKey::New);
-        let map = highlight_by_lineno(engine, syntax, &lines);
-        // Both hunks must be represented in the lineno-keyed map.
-        assert!(map.contains_key(&10), "hunk 1 missing lineno 10");
-        assert!(map.contains_key(&11), "hunk 1 missing lineno 11");
-        assert!(map.contains_key(&50), "hunk 2 missing lineno 50");
-        assert!(map.contains_key(&51), "hunk 2 missing lineno 51");
-    }
-
-    #[test]
-    fn side_cache_key_falls_back_to_content_hash_when_sha_none() {
-        // Regression guard for B2: working-tree diffs always have
-        // `new_sha = None`; the cache lookup MUST derive the same
-        // fallback key that storage uses, otherwise highlights computed
-        // under `content:...` never got served to the pane.
-        let hunk = DiffHunk {
-            old_start: 1,
-            old_lines: 1,
-            new_start: 1,
-            new_lines: 1,
-            header: "@@ -1 +1 @@".into(),
-            lines: vec![mk_line(DiffLineKind::Context, Some(1), Some(1), "hello\n")],
-        };
-        let file = mk_file("x.rs", None, None, vec![hunk]);
-        let key_new = side_cache_key(&file, SideKey::New, "Rust");
-        assert!(
-            key_new.0.starts_with("content:"),
-            "expected content fallback, got {}",
-            key_new.0
-        );
-        // Second call with identical file yields identical key — storage
-        // and lookup must agree.
-        let key_new2 = side_cache_key(&file, SideKey::New, "Rust");
-        assert_eq!(key_new, key_new2);
-    }
-
-    #[test]
-    fn side_cache_key_uses_sha_when_available() {
-        let file = mk_file("x.rs", Some("abc123"), Some("def456"), vec![]);
-        let key_new = side_cache_key(&file, SideKey::New, "Rust");
-        let key_old = side_cache_key(&file, SideKey::Old, "Rust");
-        assert_eq!(key_new.0, "def456");
-        assert_eq!(key_old.0, "abc123");
-        assert_eq!(key_new.2, SideKey::New);
-        assert_eq!(key_old.2, SideKey::Old);
-    }
-
-    #[test]
-    fn content_fallback_key_stable_for_same_content() {
-        let a = content_fallback_key("fn foo() {}\n");
-        let b = content_fallback_key("fn foo() {}\n");
-        assert_eq!(a, b);
-        let c = content_fallback_key("fn bar() {}\n");
-        assert_ne!(a, c);
-    }
-
-    #[test]
-    fn side_lines_skips_lines_without_lineno() {
-        // A malformed diff line lacking old_lineno on the old side should
-        // be skipped, not panic or emit a zero lineno.
-        let hunk = DiffHunk {
-            old_start: 1,
-            old_lines: 2,
-            new_start: 1,
-            new_lines: 2,
-            header: "@@".into(),
-            lines: vec![
-                mk_line(DiffLineKind::Context, Some(1), Some(1), "ok\n"),
-                mk_line(DiffLineKind::Context, None, None, "broken\n"),
-            ],
-        };
-        let file = mk_file("x.rs", None, None, vec![hunk]);
-        let lines = side_lines(&file, SideKey::Old);
-        assert_eq!(lines.len(), 1);
-        assert_eq!(lines[0].0, 1);
     }
 }
